Add files via upload
This file is part of an Assignment I did. It is a bash script that extracts information from a file passed as a parameter, and then it creates user accounts with their respective directories, groups, and permissions, then it sets 12 char passwords for each user. Lastly, it saves all login information to a file.